PICTUNATOR
Fixed Hatching back to where it returns an array of objects with where the line starts, ends, and width.
Now uses PurePixel.
It will also use the exact same functions with Crosshatching.
Crosshatching will use an array of the Hatching functions whose length will be equal to the number of hatches.
Since PurePixel was modified, Halftone will need to be modified too. Hidden in the meantime.
Crosshatching hidden.
Increased size of toolbar to 300px.
Will use getArrToDrawHatch() to create an SVG file.