Popup repeating only data from last clicked FIXED.
All edits were done only in the string that is passed to the to-download HTML file.
AutoDialog part fixed to include Title, Trigger, and Body as part of THIS.
AutoDialog is instantiated now only once outside of the forEach for the span links.
Now "click" listener for popup trigger links dynamically changes title and content before showing popup.

//  Will need to modify AutoDialog.js to allow for instace to be used as a template instead of leaving trigger, title, and body fixed.